Upstream iOS WebKit2 to OpenSource (part 3). https://bugs.webkit.org/show_bug.cgi?id=124803
Reviewed by Anders Carlsson and Tim Horton.



Canonical link: https://commits.webkit.org/142975@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@159724 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit2/Shared/API/Cocoa/WKRemoteObjectCoder.mm
@@ -397,7 +397,7 @@ static void decodeInvocationArguments(WKRemoteObjectDecoder *decoder, NSInvocati
         }
 
         default:
-            [NSException raise:NSInvalidArgumentException format:@"Unsupported invocation argument type '%s' for argument %zu", type, i];
+            [NSException raise:NSInvalidArgumentException format:@"Unsupported invocation argument type '%s' for argument %zu", type, (unsigned long)i];
         }
     }
 }
